pyproject.toml
[project]
name = "eodh-qgis"
version = "0.0.2"
description = "QGIS plugin for EODH"
authors = [{ name = "Dusan Figala", email = "dusan@oxidian.com" }]
license = { text = "BSD-3-Clause" }
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "pyeodh>=0.0.7",
    "qgis-plugin-ci>=2.8.10",
]

[dependency-groups]
dev = [
    "pre-commit",
    "pyright",
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "pytest-qgis",
    "pytest-watcher",
    "pytest-xdist",
    "pyqt5",
    "ruff",
    "validate-pyproject",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]

[tool.pyright]
include = ["eodh_qgis"]
exclude = ["eodh_qgis/test"]
pythonVersion = "3.13"
typeCheckingMode = "basic"
reportMissingImports = "warning"

[tool.ruff]
line-length = 119

[tool.ruff.lint]
select = [
    "E",    # Pycodestyle errors
    "W",    # Pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "FAST", # FastAPI
    "RUF",  # RUF-specific rules
    "SIM",  # Simplifications
    "B",    # Bugbear
    "ANN",  # Annotations
    "PLC",  # Pylint convention
    "PLE",  # Pylint errors
    "UP",   # pyupgrade
    "FURB", # refurb
    "PT",   # pytest
]
ignore = [
    "E501",    # line too long
    "SIM108",  # ternary operator
    "SIM103",  # return bool directly
    "SIM105",  # suppressible exception
    "SIM115",  # open-file-with-context-handler
    "B008",    # function call in default argument (incompatible with FastAPI)
    "B023",    # function uses loop variable - intentional closures in Qt signal handlers
    "ANN",     # missing type annotations - too many to add at once
    "PLC0415", # import outside top-level - intentional lazy imports in QGIS plugin
    "PT009",   # unittest-style assertions - tests use unittest via pytest-qgis
    "PT027",   # unittest-style raises - tests use unittest via pytest-qgis
    "RUF012",  # mutable class default - used in Qt classes
    "RUF059",  # unused unpacked variable
    "UP031",   # printf-string-formatting
    "FURB162", # fromisoformat-replace-z
]
